human/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pillow>=10.3.0
requests>=2.32.0
transformers>=4.41.0
accelerate>=0.30.0
opencv-python-headless>=4.9.0.80
torch>=2.1.0
torchvision>=0.16.0
git+https://github.com/facebookresearch/segment-anything.git
